+++ NEWS.md
@@ -82,6 +82,7 @@ Note: We're only listing outstanding class updates.
     * String#byteindex and String#byterindex have been added. [[Feature #13110]]
     * Update Unicode to Version 14.0.0 and Emoji Version 14.0. [[Feature #18037]]
       (also applies to Regexp)
+    * String#bytesplice has been added.  [[Feature #18598]]
 
 * Struct
     * A Struct class can also be initialized with keyword arguments